Update package.json
Colors.js package is now ongoing major changes (github.com/Marak/colors.js for references) and as of now the package 1.4.2 seems to be missing/has been removed from npm, luckly the last stable version 1.4.0 can still be used. As long as this situation doesn't deflate I suggest to keep using this version to keep the package up and running.