feat: support nesting of generators
Sub generators can include nested generators as a static property
which are resolved and installed automatically by Easy UI5 and
finally executed in a chain after the root generator and the sub
generator.

```js
export default class extends Generator {
	static displayName = "Create a new UI5 application";
	static nestedGenerators = [ "wdi5", "library:app" ];
```

Just describe the subgenerator name as you would specify it when
using Easy UI5 and if you want to address a dedicated generator
in the subgenerator, just use the namespace syntax defining the
dedicated generator with the `:`.